removed file upload thread

+++ server/api/upload/upload.js
@@ -1,6 +1,6 @@
 import { Meteor } from 'meteor/meteor';
 import bodyParser from 'body-parser';
-import { Files } from "../../../shared/collections/files";
+import { Files } from "/shared/collections/files";
 import { app } from "../express";
 
 const _multer = require('multer');
@@ -38,33 +38,49 @@ app.post('/api/upload', [
 app.put('/api/upload', [
 	_multerInstance.single('file')
 ], Meteor.bindEnvironment((request, response, next) => {
-	if (request.file !== undefined) {
-		response.writeHead(201, {
-			'Content-Type': 'application/json'
-		});
-		response.end(JSON.stringify(Meteor.wrapAsync(async (callback) => {
-			callback(null, await (() => {
-				return new Promise((resolve, reject) => {
-					_fs.stat(request.file.path, function (_statError, _statData) {
-						const _addFileMeta = {
-							fileName: request.file.originalname,
-							type: request.file.mimetype,
-							size: request.file.size
-						};
-						_fs.readFile(request.file.path, function (_readError, _readData) {
+		if (request.file !== undefined) {
+			_fs.stat(request.file.path, function (_statError, _statData) {
+				if (_statError === null) {
+					const _addFileMeta = {
+						fileName: request.file.originalname,
+						type: request.file.mimetype,
+						size: request.file.size
+					};
+					_fs.readFile(request.file.path, function (_readError, _readData) {
+						if (_readError === null) {
 							Files.write(_readData, _addFileMeta, function (_uploadError, _uploadData) {
-								resolve(_uploadData);
+								if (_uploadError !== undefined) {
+									response.writeHead(201, {
+										'Content-Type': 'application/json'
+									});
+									response.end(JSON.stringify(_uploadData));
+								} else {
+									response.writeHead(500, {
+										'Content-Type': 'application/json'
+									});
+									response.end();
+								}
 								_fs.unlink(request.file.path);
 							}, true);
-						});
+						} else {
+							response.writeHead(500, {
+								'Content-Type': 'application/json'
+							});
+						}
 					});
-				});
-			})());
-		})()));
-	} else {
-		response.writeHead(400, {
-			'Content-Type': 'application/json'
-		});
-		response.end();
+				} else {
+					response.writeHead(500, {
+						'Content-Type': 'application/json'
+					});
+					response.end();
+				}
+			});
+		}
+		else {
+			response.writeHead(400, {
+				'Content-Type': 'application/json'
+			});
+			response.end();
+		}
 	}
-}));
+));
